requirements.txt
alembic==1.4.2
Click==7.0
dnspython==1.16.0
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.2
MarkupSafe==1.1.1
mongoengine==0.19.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
pymongo==3.10.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.13.0
SQLAlchemy==1.3.16
Unidecode==1.1.1
Werkzeug==0.16.0
WTForms==2.2.1
